Enable possibility of Aarch64 build
Tizen toolchain team in Samsung company is planning to build Tizen
for new arm 64 bit architecture - Aarch64. But package
"haptic-module-tizen" cannot be built for this architecture beacase
cmake build scripts do not accept aarch64 as appropriate
architecture for the build.

That's why it has been added to the list of architectures for
which the package can be built.

Change-Id: Ie915f959110bb906648ebbf8a559756cd44d6e92
Signed-off-by: Ilya Palachev <i.palachev@samsung.com>